pyproject.toml
[project]
name = "kineticconstructs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "astrapy>=2.0.1",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
